Add voucher management views: create, edit, and index pages
- Implemented the create voucher form with fields for code, description, discount type, value, minimum purchase, quota, expiration date, and settings.
- Developed the edit voucher page to allow modifications of existing vouchers, displaying current values and usage statistics.
- Created an index page to list all vouchers with filtering options, displaying their status, type, value, and actions for editing or deleting.

+++ database/migrations/2024_08_26_024545_create_products_table.php
@@ -100,7 +100,7 @@ return new class extends Migration
             // Contoh:
             // 10 = 10%
             // 5000 = Rp5.000
-            $table->bigInteger('discount_value')
+            $table->decimal('discount_value', 10, 2)
                 ->nullable();
 
             // Stok produk
